pyproject.toml
[project]
name = "hivemc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=46.0.6",
    "flask>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
    "pymysql>=1.1.2",
    "pyyaml>=6.0.3",
]
